Added new \Common\Core\Model class to use inheritance for Backend, Frontend and Api Init classes and moved all common methods to that class.
